Revert "Create UI for add-maintainer"
