+++ src/main/java/com/google/sps/servlets/AddMaintainerServlet.java
@@ -53,13 +53,13 @@ public class AddMaintainerServlet extends HttpServlet {
     // User can add another Maintainer only if they are a Maintainer.
     String newMaintainerEmail = request.getParameter("userEmail");
 
-    Entity entity = GivrUser.getUserFromDatastoreWithProperty("userEmail", newMaintainerEmail);
+    Entity entity = GivrUser.getUserFromDatastoreWithProperty("userEmail", currUser.getUserEmail());
     if (entity != null) {
-      changeMaintainerStatus(newMaintainerEmail);
+      changeMaintainerStatus(newMaintainerEmail, currUser.getUserId());
     } else {
       addNewMaintainerToDatastore(newMaintainerEmail);
     }
-    response.sendRedirect("/index.html");
+    response.setStatus(HttpServletResponse.SC_OK);
   }
 
   public void addNewMaintainerToDatastore(String email) {
@@ -73,10 +73,11 @@ public class AddMaintainerServlet extends HttpServlet {
     datastore.put(newUserEntity);
   }
 
-  public void changeMaintainerStatus(String email) {
+  public void changeMaintainerStatus(String email, String userId) {
     boolean isMaintainer = true;
     Map<String, Object> propertyNamesAndValuesToUpdate = new HashMap<String, Object>();
     propertyNamesAndValuesToUpdate.put("isMaintainer", isMaintainer);
+    propertyNamesAndValuesToUpdate.put("userId", userId);
     GivrUser.updateUserInDatastore("userEmail", email, propertyNamesAndValuesToUpdate);
   }
 }
