+++ src/main/webapp/index.html
@@ -7,6 +7,7 @@
 	  <script src="user.js"></script>
 	  <script src="organization.js"></script>
 	  <script src="search_area.js"></script>
+    <script src="maintainer.js"></script>
   </head>
   
   <body>
