+++ src/main/webapp/style.css
@@ -45,7 +45,6 @@
   margin: 40px 50px 20px 50px;
 }
 
-#maintainer-email-entry,
 #zipcode-entry {
   width: 20%;
   margin: 8px 0;
@@ -79,7 +78,7 @@
 }
 
 #filter-datalist {
-  float: right;
+    float: right;
 }
 
 #organization-list {
@@ -119,7 +118,7 @@
 }
 
 .organization-popup-name {
-  margin: 2% auto;
+    margin: 2% auto;
 }
 
 .organization-popup-phone,
@@ -129,7 +128,7 @@
 .organization-popup-email,
 .organization-popup-url-link,
 .organization-popup-description {
-  margin: 2% 0;
+    margin: 2% 0;
 }
 
 .hide-popup,
@@ -148,10 +147,6 @@
   cursor: pointer;
 }
 
-#add-maintainer-label {
-  cursor: pointer;
-}
-
 #registration-form-header{ 
   display: flex;
   margin: 1% auto 2% auto;
@@ -160,8 +155,8 @@
 }
 
 #registration-form-area {
-  width: 60%;
-  margin: 0 auto;
+    width: 60%;
+    margin: 0 auto;
 }
 
 .registration-entry {
@@ -204,10 +199,6 @@
   border: 1px solid #888;
   width: 70%;
 }
-
-#add-maintainer-email-form {
-  display: flex;
-  flex-flow: row nowrap;
 
 .day-option-area {
   margin: 5px 0px 5px 0px;
@@ -270,28 +261,3 @@
   font-size: 20px;
 }
 
-.add-maintainer-popup-modal-background {
-  display: flex;
-  justify-content: center;
-  z-index: 1;
-  background-color: rgba(0,0,0,0.4);
-  position: absolute;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-}
-
-#add-maintainer-popup-modal {
-  margin: auto;
-  padding: 10px;
-  background-color: white;
-  background-clip: content-box;
-  box-shadow: inset 0 0 0 10px white;
-  align-items: center;
-  justify-content: center;
-  display: flex;
-  width: 30%;
-  height: 5%;
-  font-family: 'Courier New', Courier, monospace;
-  color: #666666;
-}
